server/requirements.txt
django==1.9.13
djangorestframework==3.4
django-filter==2.1.0
django-cors-headers==1.1.0
dj_database_url==0.4.0
envtpl==0.6.0
requests==2.9.1
psycopg2==2.6.1
pandas==0.17.1
numpy
uwsgi==2.0.18
path.py==7.2